Strangler extraction #2: online API layer -> zxnu_api.py (+unit tests, star-import tripwire)
- New zxnu_api.py (1,463 lines): the module-level online catalogue API
  layer moved verbatim out of zx-next-unite.py — shared HTTP retry
  helpers, GetIt fetch + caret-format parsers, ZXDB fetch/pick/search/
  detail parsers + website-URL builder, zxArt safe-URL encoder, fetchers,
  prod/picture parsers, prefix search and the thread-safe name-resolution
  caches. Pure Python, no Qt. zx-next-unite.py drops 1,430 lines (~24.4k
  now; ~1.8k down across the two extractions).
- Star-imported back so all historical names keep working, PLUS an
  explicit import of the 12 underscore-prefixed helpers the monolith
  still references: 'import *' silently skips private names, which broke
  every gallery's image loading mid-branch (_gif_fetch_bytes ->
  _http_fetch_bytes_with_retry NameError, swallowed by the worker's
  error path). Fixed and guarded: tests/test_api_parsers.py carries a
  static tripwire asserting the explicit list stays complete.
- tests/test_api_parsers.py: 30 fast checks (0.2s, no network/Qt) over
  the caret format, both ZXDB envelope shapes, URL builders, download
  filter, Cyrillic URL encoding, zxArt prod/picture parsers + the
  tripwire; wired into tests/run_all.py. CLAUDE.md source map updated.
