+++ tests/run_all.py
@@ -20,6 +20,7 @@ HERE = os.path.dirname(os.path.abspath(__file__))
 
 SUITES = [
     # (file, timeout seconds, required import or None)
+    ("test_api_parsers.py",     120, None),
     ("test_listen.py",          120, None),
     ("test_remote_listen.py",   120, None),
     ("test_http_bridge.py",     240, "flask"),
